requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
courlan==1.3.1
dateparser==1.2.0
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
Faker==30.1.0
Flask==3.0.3
flask_migrate==4.0.7
flask_sqlalchemy==3.1.1
Flask-Session==0.8.0 
flask_cors==4.0.0
frozenlist==1.4.1
gevent>=1.4.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
git+https://github.com/google-gemini/generative-ai-python@imagen
graphviz==0.20.3
gunicorn==23.0.0
h11==0.14.0
holidays==0.60
htmldate==1.9.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imgurpython==1.1.7
importlib_metadata==8.5.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jusText==3.0.1
lxml==5.3.0
lxml_html_clean==0.2.2
MarkupSafe==3.0.1
multidict==6.1.0
oauthlib==3.2.2
openai==1.50.2
openpyxl==3.0.9
outcome==1.3.0.post0
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
psycopg2
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
PyDrive==1.3.1
pyparsing==3.1.4
pyshorteners==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tiktoken==0.8.0
tld==0.13
tqdm==4.66.5
trafilatura==1.12.2
trio==0.26.2
trio-websocket==0.11.1
twitchio==2.10.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
